Resources/cae/requirements.txt
bagoftools
numpy
Pillow
PyYAML
scikit_image
torch
torchvision
